requirements.txt
streamlit 
python-dotenv
bs4
google-generativeai